Add refactoring of Landing and Dashboard components
