+++ src/components/App/Landing/middleware/change-order-view.js
@@ -1,4 +1,4 @@
 module.exports = function changeOrderView(sType) {
   const sT = sType || 'all';
-  this.setStateVal({ viewBy: sT });
+  this.props.setStateVal({ viewBy: sT });
 };
