+++ src/components/App/Landing/middleware/show-other-form.js
@@ -1,0 +1,5 @@
+module.exports = function showOtherForm() {
+  this.props.setStateVal({
+    showStockOrderForm: !this.props.showStockOrderForm
+  });
+};
