--- src/components/App/middleware/show-other-form.js
@@ -1,3 +1,0 @@
-module.exports = function showOtherForm() {
-  this.setStateVal({ showStockOrderForm: !this.state.showStockOrderForm });
-};
