+++ src/sass/components/Dashboard.scss
@@ -24,6 +24,9 @@
     margin: $Dashboard_padding;
     margin: auto;
     position: relative;
+    button {
+
+    }
   }
 }
 
@@ -81,6 +84,14 @@
     background-position: center;
     width: calc(26% - (#{$button_border_size}*2) - 0.7em);
     margin-right: 1em;
+  }
+  &.logout {
+    font-size: 1em;
+    text-align: center;
+    span {
+      width: 100%;
+      display: block;
+    }
   }
 }
