+++ src/sass/components/Landing.scss
@@ -32,6 +32,7 @@
         color: $c6;
         border-bottom: solid 1px $c6;
         font-weight: 300;
+        text-transform: capitalize;
         &:hover {
           color: $primary_blue;
           border-bottom: solid 1px $primary_blue;
